Add 384 well selecting quadrants test for well_from_shape
well_from_shape was not returning correct wells for index beyond one.
For example, for a 384 well with origin well of A2 and shape of 96 well
to select, it should return the 2nd quadrant, however, the current code
returns the 1st quadrant wells. This commit adds the tests and shows
the test are failing.