build: pre-commit infrastructure and update travis (#243)
Add pre-commit infrastructure to help with basic lint tasks. This is in
preparation of adding an autoformatter for the repo.

As part of this, the travis file is reorganized to help with clearer
delineation of jobs. The underlying dist is also bumped to `bionic`
since there aren't any explicit dependencies on `xenial` and `bionic`
is stable.

Documention is also updated, with the Contributing section updated with
more explicit testing steps and the new environment and lint setup.

This includes all the fixes from the pre-commit hooks added.