docs: Bump documentation dependencies
As we're preparing for a major release, this is an opportunity to
upgrade some of our base dependencies.

This PR bumps our base Sphinx version to >=2.4, and also bumps our
other documentation related dependencies to their newest versions.

Notably, this brings in many bugfixes and some nicer UI elements such as
a nicer search bar.

The full changelog is available here:
https://www.sphinx-doc.org/en/master/changes.html#release-2-4-4-released-mar-05-2020

Verified that the docs build fine.