feat: Bump Pint to 0.9
As we're preparing for a major release, this is an opportunity to
upgrade some of our base dependencies.

This upgrades the base Pint version to 0.9 which brings some bugfixes
and better interop with other Python libraries.

See https://github.com/hgrecco/pint/blob/master/CHANGES#L259

Ideally we'll bump to 0.11, but there are some incompat issues which
we'll need to investigate.